02_Career_Mentor_Agent/pyproject.toml
[project]
name = "02-career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.3.0",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
